refactor(VLEComponent): Extract NodeClickLockedService
