+++ src/assets/wise5/vle/vle.component.spec.ts
@@ -31,6 +31,7 @@ import { InitializeVLEService } from '../services/initializeVLEService';
 import { StudentTeacherCommonServicesModule } from '../../../app/student-teacher-common-services.module';
 import { PauseScreenService } from '../services/pauseScreenService';
 import { StudentNotificationService } from '../services/studentNotificationService';
+import { NodeClickLockedService } from '../services/nodeClickLockedService';
 
 let component: VLEComponent;
 let fixture: ComponentFixture<VLEComponent>;
@@ -71,6 +72,7 @@ describe('VLEComponent', () => {
       ],
       providers: [
         InitializeVLEService,
+        NodeClickLockedService,
         PauseScreenService,
         ProjectService,
         StudentNotificationService,
